backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.2
python-dotenv>=1.0.0
numpy>=1.25.2,<2.0.0
scikit-learn>=1.3.1
requests>=2.31.0
google-generativeai>=0.3.1
python-multipart>=0.0.6
sentence-transformers>=2.2.2
pandas>=2.1.1
aiofiles>=23.2.1
httpx>=0.25.0
chromadb